+++ CHANGELOG.md
@@ -7,7 +7,7 @@
 - `URLSearchParams::sort()` now correctly sorts by code units instead of trying to fake it using code points.
 - Null bytes are now percent encoded in fragments instead of being ignored per [whatwg/url#440](https://github.com/whatwg/url/issues/440).
 - Domain names of URLs with special schemes can no longer be an empty string per [whatwg/url#492](https://github.com/whatwg/url/pull/492) and [whatwg/url#497](https://github.com/whatwg/url/pull/497).
-- Host names now also forbid the use of the `&lt;`, `&gt;`, and `^` code points per [whatwg/url#458](https://github.com/whatwg/url/issues/458).
+- Host names now also forbid the use of the `<`, `>`, and `^` code points per [whatwg/url#458](https://github.com/whatwg/url/issues/458).
 
 ### Fixed